# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Paytm Smart-Stock AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InventoryItemStatus =
  (typeof InventoryItemStatus)[keyof typeof InventoryItemStatus];

export const InventoryItemStatus = {
  safe: "safe",
  low: "low",
  critical: "critical",
} as const;

export interface InventoryItem {
  id: string;
  name: string;
  currentStock: number;
  reorderThreshold: number;
  maxStock: number;
  price: number;
  unit: string;
  status: InventoryItemStatus;
}

export interface UpdateInventoryRequest {
  stock: number;
}

export interface Transaction {
  id: string;
  productId: string;
  productName: string;
  quantity: number;
  price: number;
  total: number;
  timestamp: string;
}

export interface CreateTransactionRequest {
  productId: string;
  quantity: number;
}

export interface StockAlert {
  triggered: boolean;
  productId?: string;
  productName?: string;
  currentStock?: number;
  reorderQty?: number;
  orderCost?: number;
  walletBalance?: number;
  needsLoan?: boolean;
}

export interface TransactionResult {
  transaction: Transaction;
  updatedInventory: InventoryItem;
  alert?: StockAlert;
}

export interface Prediction {
  productId: string;
  productName: string;
  burnRate: number;
  predictedDemand: number;
  weatherMultiplier: number;
  confidence: string;
}

export interface Wallet {
  balance: number;
  currency: string;
}

export interface LoanRequest {
  amount: number;
}

export type GetPredictionsParams = {
  dayType?: GetPredictionsDayType;
};

export type GetPredictionsDayType =
  (typeof GetPredictionsDayType)[keyof typeof GetPredictionsDayType];

export const GetPredictionsDayType = {
  normal: "normal",
  rainy: "rainy",
  festival: "festival",
} as const;