# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Paytm Smart-Stock AI API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItemStatus } from "./inventoryItemStatus";

export interface InventoryItem {
  id: string;
  name: string;
  currentStock: number;
  reorderThreshold: number;
  maxStock: number;
  price: number;
  unit: string;
  status: InventoryItemStatus;
}