# lib/api-zod/src/generated/types/stockAlert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Paytm Smart-Stock AI API
 * OpenAPI spec version: 0.1.0
 */

export interface StockAlert {
  triggered: boolean;
  productId?: string;
  productName?: string;
  currentStock?: number;
  reorderQty?: number;
  orderCost?: number;
  walletBalance?: number;
  needsLoan?: boolean;
}