# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Paytm Smart-Stock AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Transaction {
  id: string;
  productId: string;
  productName: string;
  quantity: number;
  price: number;
  total: number;
  timestamp: string;
}