# lib/api-zod/src/generated/types/transactionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Paytm Smart-Stock AI API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItem } from "./inventoryItem";
import type { StockAlert } from "./stockAlert";
import type { Transaction } from "./transaction";

export interface TransactionResult {
  transaction: Transaction;
  updatedInventory: InventoryItem;
  alert?: StockAlert;
}