# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Paytm Smart-Stock AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Wallet {
  balance: number;
  currency: string;
}